Fix copyBibtexButton event listener by removing remove duplicate code.
Also improve error handling when copying text. We now ignore errors that would raise when you display the options.html page -- whose url is extensions:// and the script doesn't have permission to copy text from there. But we don't need to highlight and copy text from there anyway, so it's safe to ignore.